Split frontend logic from backend logic.
This is a very big refactoring which incorporated a lot of design
changes in order to accomplish the frontend vs. backend separation.

Previously, frontend & backend were part of a single monolithic 'Yavide'
(https://github.com/JBakamovic/yavide) repository which is a heavily
modded Vim.

This commit wraps a frontend logic into a plain Vim plugin which can be
used with whatever Vim configuration and not only 'Yavide'.

Wrote a completely new documentation.

A lot of stuff I didn't mention ... See the remove_backend branch log
history for more details.