Fix performance issues caused by semantic syntax highlighting service.
Generate and apply that many syntax rules as there are visible lines in
the window. This relieves the Vim rendering engine from not applying syntax
rules to the lines which are not visible. For big files this a huge win-win.

Now we have smooth scrolling again.

For this to implement reliably we would need some kind of 'ViewportChanged'
event in Vim which we unfortunatelly don't have. Closest we can get is
'CursorHold(I)' event.

It is not perfect byt however it does feel _much_ better than before.