Fix a bug in modifications-handler.
Seems like that this routine is the main cause that we end-up serializing
files unnecessary (when we don't expect them to). This in turn triggers
`cxxd` backend to re-parse the given TU needlessly, which degrades overall
performance.

'changedtick' seems to register movements in the buffer such as moving the
cursor along the stack (previous/next position; alt-left, alt-right).
This is most certainly something which we don't want to trigger re-parsing
with. Current 'fix' is to check if buffer has been really changed and then
move onto the 'changedtick' condition. Not sure if that will be the final
fix but we'll see. 'undo' comes to my mind which was the reason to use
'changedtick' (to detect those actions and re-parse the TU to get a fresh
syntax-highlighting).