Make use of new Vim terminal feature to show the whole build process.
The whole process of building the executable is now displayed at the
bottom of the screen in the full width of the window. Once done, build
output will be, as previously, fed into the qf window so one can browse
the output and jump to errors/warnings interactively. Terminal is
automatically closed once the qf is populated.