Implement the frontend part for code completion.
It is not completely polished, some more vimscript magic foo is needed,
but most of the time works ok. It isn't instant as I would like it to be,
there's a lot of glue-layer processing written in Python and Vimscript.

Given that I have no particular incentive as of now to do a total rewrite
in some other compiled language with the goal of speeding it up, and
given that I've been running and using this code from a feature branch
for so much long, I decided to merge this feature as is. Will see where
it goes.