Address a somewhat racey condition.
On BufWritePost event two cxxd services will be triggered roughly at
the same time:
 1. indexer - to keep the indexing symbol database up to date
 2. clang-format - to apply the source code formatting rules

Due to the cxxd architecture which decouples services one from each other
and enabling them to run in parallel, it is possible to run into a situation
where indexer will manage to update the indexing symbol database _before_
the clang-format service will apply the formatting rules.

Because clang-formatting will in many cases jumble the symbols from one
column to another or even one row to another, this will have a consequence
that our indexing symbol database will now have stale information about
the symbols in the code.

To easily address this without introducing complexity, we can:
 1. Disable indexer running on BufWritePost event
 2. Have clang-format callback trigger the indexer implicitly

This approach will not suffer from the above mentioned situation, e.g.
our indexing symbol database will be always up to date, but it will
heavily rely on the fact that clang-format will automatically run on each
file save. This happens to be the case right now, because that's how it
is implemented in the cxxd-vim frontend, but it is a subject to change
and therefore it may break the indexing in subtle way.

To address the robustness, we can leave the BufWritePost event for indexer
while still having the clang-format triggering the indexer implicitly but
at the cost that now we will have the indexer always running twice in a
row.

This patch favors less robust solution rather than having indexer wastefully
running twice. Proper solution would require introducing additional logic
to synchronize between these two types of events, e.g. have a frontend
implementing the logic to deduce whether the indexer needs to send a
request on BufWritePost or not. This would depend on whether the
clang-format service is configured to be run automatically or not.