Add utility to help convert existing colorschemes into cxxd compatible ones.
cxxd has a semantic understanding of the C and C++ code. Because of that
it can generate Vim colorscheme tags/groups at much more finer level than
what vanilla Vim colorscheme will support. This is called semantic syntax
highlighting and is the reason why I have implemented yaflandia
colorscheme from the very early beginning.

This tool is envisioned to help convert your favorite Vim colorscheme
into a format so that cxxd-vim can take advantage of semantic syntax
highlighting support built into a cxxd.

What will it basically do is that it will link special syntax highlighting
groups generated by the cxxd to the groups defined by Vim as "preferred"
so it shouldn't be destructive by any means.

Preferred groups are only handful and source-code wise those include
only [Identifier, Statement, PreProc, Constant, Type]. This is why
conversion is rather conservative and it may or may not end up with the
best results. It mostly depends on how well and detailed the given
colorscheme is implemented and/or tweaked towards the C and C++ syntax.