Implement the counterpart frontend for the new cxxd disassembly service.
TL;DR cxxd disassembly service implements a Godbolt-like functionality
but extended beyond what the Godbolt does. More details are contained
in the cxxd repo, in #5b0253dfc051b5f1c864c4207fbc401c004dd084.

This is a frontend part for that feature. For better experience, it is
recommended to install the enhanced version of ASM syntax highlighting,
e.g. https://github.com/Shirk/vim-gas.

Usage is quite simple:
 (0) Start the cxxd server against your code, e.g.
       CxxdStart development/my_code RelWithDebInfo
 (1) Build your code
 (2) Open some source-code file of interest
 (3) Use CxxdDisassemblyPickTarget to get a popup-menu to select the
     target you want to inspect this function (symbol) from. This is
     important because symbol can exist in multiple different targets,
     e.g. unit-tests, benchmarks, application executable, etc.
 (4) Position the cursor on some interesting function
 (5) Use CxxdDisassemblyPickSymbol to get a popup-menu to select the
     (demangled) symbol from. Sometimes number of items in the popup-list
     will be > 1 and that is because the symbol does not exist in given
     target (e.g. it was optimized out). However, there will be a list
     of approximate symbols to pick from.
 (6) On left hand side, disassembled target will be displayed at the
     location of a symbol selected in previous step.
 (7) Mouse-hover over ASM instructions will give you documentation for
     that specific instruction.

Popups can be scrolled through with <C-n>, <C-p>, <C-t>, ..., <PageUp>
etc keyboard shortcuts as expected.

If no candidates are found, be it target (not quite probable) or
symbol (probable), a warning at the bottom of the screen will be given
and no (empty) popup-menu will be presented.