Migrate to async jobs and JSON streaming.
Originally, engine was developed in the pre-Vim8/Neovim era, at the time
when no Vim8 async jobs existed so pretty much the only way to implement
async processing was by relying on clientserver architecture, which in
itself was limiting - plain vim normally does not come with clientserver
support compiled in, neither neovim does.

So, this commit is a long overdue change which switches from clientserver
architecture to async-job architecture, which uses JSON-formatted requests
to communicate with the backend, instead of previously, multiprocessing.queue.